+++ front-end/enbalde/src/app/pages/login/login.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, ElementRef } from '@angular/core';
 import { FormBuilder, FormGroup, RangeValueAccessor, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { AuthService } from 'src/app/services/auth.service';
 import { TokenUsuario, UsuariosService } from 'src/app/services/usuarios.service';
 
 @Component({
@@ -14,7 +15,8 @@ export class LoginComponent implements OnInit {
   loginForm!: FormGroup;
   usuario = {user: '', password: ''};
 
-  constructor(private fb: FormBuilder, private router: Router, private elementRef: ElementRef, private usuariosService: UsuariosService) {}
+  constructor(private fb: FormBuilder, private router: Router, private elementRef: ElementRef, private usuariosService: UsuariosService, private authService: AuthService) {
+  }
 
   ngOnInit(): void {
       this.loginForm = this.fb.group({
@@ -27,13 +29,13 @@ export class LoginComponent implements OnInit {
   get password() { return this.loginForm.get('password'); }
 
   onSubmit(value: any) {
-    this.usuariosService.login(value.user, value.password)
+    this.authService.login(value.user, value.password)
       .subscribe(resultado => {
         let tokenUsuario = resultado.data as TokenUsuario;
 
         localStorage.setItem('accessToken', `${tokenUsuario.accessToken.acceso}`);
         localStorage.setItem('usuarioActual', JSON.stringify(tokenUsuario.usuarioActual));
-        this.usuariosService.UsuarioIngresando.emit(tokenUsuario.usuarioActual);
+        this.authService.autenticadoComo(tokenUsuario.usuarioActual);
         this.router.navigate(['/']);
         this.elementRef.nativeElement.ownerDocument.documentElement.scrollTop = 0;
       });
