+++ front-end/enbalde/src/app/shared/header/header.component.ts
@@ -3,6 +3,7 @@ import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
 import { ResultadoApi } from 'src/app/models/modelo.resultado';
 import { Usuario, TipoUsuario } from 'src/app/models/modelo.usuario';
+import { AuthService } from 'src/app/services/auth.service';
 import { ProductosService } from 'src/app/services/productos.service';
 import { UsuariosService } from 'src/app/services/usuarios.service';
 
@@ -10,7 +11,7 @@ import { UsuariosService } from 'src/app/services/usuarios.service';
   selector: 'app-header',
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.css'],
-  providers: [ ProductosService, UsuariosService ]
+  providers: [ ProductosService, UsuariosService, AuthService ]
 })
 
 export class HeaderComponent {
@@ -19,23 +20,22 @@ export class HeaderComponent {
   buscarResults!: any[];
   showResults: boolean = false
 
-  constructor (private productosService: ProductosService, private usuariosService: UsuariosService, private router: Router) {
-    this.usuariosService.UsuarioIngresando
-      .subscribe((u: Usuario) => {
-        this.usuario = u;
+  constructor (private productosService: ProductosService, private usuariosService: UsuariosService, private authService: AuthService, private router: Router) {
+    this.authService.autenticado
+      .subscribe((auth: boolean) => {
+        if (auth) {
+          this.usuario = this.authService.usuario;
+        }
       });
   }
 
   ngOnInit(): void {
     let token = localStorage.getItem('accessToken');
     let usuarioActual = localStorage.getItem('usuarioActual');
-    console.log(usuarioActual)
     if (usuarioActual) {
-      console.log(1)
       this.usuario = JSON.parse(usuarioActual) as Usuario;
     }
     else {
-      console.log(2)
       this.usuario = undefined;
     }
   }
@@ -52,12 +52,12 @@ export class HeaderComponent {
   }
 
   logout() {
-    this.usuariosService.logout()
+    this.authService.logout()
       .subscribe((resultado: ResultadoApi) => {
         if (resultado.status == HttpStatusCode.Ok) {
           localStorage.removeItem('accessToken');
           localStorage.removeItem('usuarioActual')
-          this.usuariosService.UsuarioIngresando.emit(null);
+          this.usuario = undefined;
           this.router.navigate((['/']));
         }
       });
